Calculate the arithmetic mean
This is a C language program that calculates the arithmetic mean of three numbers entered by the user. First, the program prompts the user to input the three numbers. Then, it calculates the arithmetic mean of the three numbers (the sum of the three numbers divided by three) and stores the result in the variable "media" (which means "average" in Spanish). Finally, the program prints the result of the arithmetic mean on the screen.